refactor(server): move route validation helpers into package
